+++ apps/cli/src/services/portfolio/index.ts
@@ -1,11 +1,14 @@
-import { getDefaultUnit, getParsedAmount } from '@cypherock/coin-support-utils';
-import { coinList } from '@cypherock/coins';
+import {
+  getAsset,
+  getDefaultUnit,
+  getParsedAmount,
+} from '@cypherock/coin-support-utils';
 import { getCoinAllocations } from '@cypherock/cysync-core-services';
 import { BigNumber } from '@cypherock/cysync-utils';
 import { IDatabase } from '@cypherock/db-interfaces';
 
-export const showPortfolio = async (db: IDatabase) => {
-  const allocations = await getCoinAllocations({ db });
+export const showPortfolio = async (db: IDatabase, currency: string) => {
+  const allocations = await getCoinAllocations({ db, currency });
 
   const displayTable = allocations
     .sort((a, b) => b.percentage - a.percentage)
@@ -13,14 +16,16 @@ export const showPortfolio = async (db: IDatabase) => {
       const { amount, unit } = getParsedAmount({
         coinId: a.parentAssetId,
         assetId: a.assetId,
-        unitAbbr: getDefaultUnit(a.assetId).abbr,
+        unitAbbr: getDefaultUnit(a.parentAssetId, a.assetId).abbr,
         amount: a.balance,
       });
 
+      const asset = getAsset(a.parentAssetId, a.assetId);
+
       return {
-        name: coinList[a.assetId].name,
+        name: asset?.name,
         balance: `${amount} ${unit.abbr}`,
-        value: `${new BigNumber(a.value).toFixed(2)} USD`,
+        value: `${new BigNumber(a.value).toFixed(2)}`,
         percentage: `${a.percentage.toFixed(2)}%`,
       };
     });