+++ packages/cysync-core/src/actions/account/priceHistoriesSync.ts
@@ -7,7 +7,13 @@ import { Observer } from 'rxjs';
 
 import { getDB } from '~/utils';
 
-export const syncPriceHistories = ({ families }: { families: string[] }) =>
+export const syncPriceHistories = ({
+  families,
+  currency,
+}: {
+  families: string[];
+  currency: string;
+}) =>
   new Promise<void>(resolve => {
     const observer: Observer<ISyncPriceHistoriesEvent> = {
       error: () => {
@@ -23,8 +29,9 @@ export const syncPriceHistories = ({ families }: { families: string[] }) =>
     syncPriceHistoriesCore({
       db: getDB(),
       families,
+      currency,
     }).subscribe(observer);
   });
 
-export const syncAllPriceHistories = () =>
-  syncPriceHistories({ families: Object.values(coinFamiliesMap) });
+export const syncAllPriceHistories = (currency: string) =>
+  syncPriceHistories({ families: Object.values(coinFamiliesMap), currency });