+++ packages/cysync-core/src/dialogs/Send/context/index.tsx
@@ -63,6 +63,7 @@ import { LoaderDialog } from '~/components';
 import {
   WalletConnectCallRequestMethodMap,
   deviceLock,
+  useCurrency,
   useDevice,
   useWalletConnect,
 } from '~/context';
@@ -76,8 +77,8 @@ import {
 } from '~/hooks';
 import {
   closeDialog,
+  selectCurrentCurrencyPriceInfos,
   selectLanguage,
-  selectPriceInfos,
   useAppDispatch,
   useAppSelector,
 } from '~/store';
@@ -203,7 +204,10 @@ export const SendDialogProvider: FC<SendDialogContextProviderProps> = ({
 }) => {
   const lang = useAppSelector(selectLanguage);
   const dispatch = useAppDispatch();
-  const { priceInfos } = useAppSelector(selectPriceInfos);
+  const { currentCurrency } = useCurrency();
+  const priceInfos = useAppSelector(state =>
+    selectCurrentCurrencyPriceInfos(state, currentCurrency),
+  );
   const deviceRequiredDialogsMap: Record<number, number[] | undefined> =
     useMemo(
       () => ({
@@ -721,9 +725,7 @@ export const SendDialogProvider: FC<SendDialogContextProviderProps> = ({
 
   const priceConverter = (val: string, invert?: boolean) => {
     const coinPrice = priceInfos.find(
-      p =>
-        p.assetId === selectedAccount?.assetId &&
-        p.currency.toLowerCase() === 'usd',
+      p => p.assetId === selectedAccount?.assetId,
     );
 
     if (!coinPrice) return '';
@@ -736,7 +738,7 @@ export const SendDialogProvider: FC<SendDialogContextProviderProps> = ({
     if (result.isNaN()) return '';
     return invert
       ? formatDisplayAmount(result).complete
-      : formatDisplayPrice(result);
+      : formatDisplayPrice(result, currentCurrency);
   };
 
   const updateUserInputs = (count: number) => {