+++ packages/cysync-core/src/hooks/useTransactions.tsx
@@ -30,11 +30,12 @@ import React, { useCallback, useEffect, useState } from 'react';
 
 import { openHistoryDialog } from '~/actions';
 import { CoinIcon } from '~/components';
+import { useCurrency } from '~/context';
 import { useStateToRef, useWindowSize } from '~/hooks';
 import {
   selectDiscreetMode,
   selectLanguage,
-  selectPriceInfos,
+  selectCurrentCurrencyPriceInfos,
   selectTransactions,
   selectUnHiddenAccounts,
   selectWallets,
@@ -129,16 +130,18 @@ const selector = createSelector(
     selectWallets,
     selectUnHiddenAccounts,
     selectTransactions,
-    selectPriceInfos,
+    selectCurrentCurrencyPriceInfos,
     selectDiscreetMode,
+    (state, currency: string) => currency,
   ],
   (
     lang,
     { wallets },
     { accounts },
     { transactions },
-    { priceInfos },
+    priceInfos,
     { active: isDiscreetMode },
+    currency,
   ) => ({
     lang,
     wallets,
@@ -146,6 +149,7 @@ const selector = createSelector(
     transactions,
     priceInfos,
     isDiscreetMode,
+    currency,
   }),
 );
 
@@ -178,18 +182,20 @@ const getDisplayValues = (
   transaction: ITransaction,
   priceInfos: IPriceInfo[],
   isDiscreetMode: boolean,
+  currency: string,
 ) => {
-  let displayValue = '$0.00';
+  let displayValue = formatDisplayPrice('0.00', currency);
   let value = '0.00';
-  let displayFeeValue = '$0.00';
+  let displayFeeValue = formatDisplayPrice('0.00', currency);
   const coinPrice = priceInfos.find(
     p =>
       p.assetId === transaction.parentAssetId &&
-      p.currency.toLowerCase() === 'usd',
+      p.currency.toLowerCase() === currency.toLowerCase(),
   );
   const assetPrice = priceInfos.find(
     p =>
-      p.assetId === transaction.assetId && p.currency.toLowerCase() === 'usd',
+      p.assetId === transaction.assetId &&
+      p.currency.toLowerCase() === currency.toLowerCase(),
   );
 
   if (coinPrice) {
@@ -202,7 +208,7 @@ const getDisplayValues = (
     const feeValue = new BigNumber(feeInDefaultUnit.amount).multipliedBy(
       coinPrice.latestPrice,
     );
-    displayFeeValue = `$${formatDisplayPrice(feeValue)}`;
+    displayFeeValue = formatDisplayPrice(feeValue, currency);
   }
 
   if (assetPrice) {
@@ -219,10 +225,10 @@ const getDisplayValues = (
       new BigNumber(amountInDefaultUnit.amount).multipliedBy(
         assetPrice.latestPrice,
       ),
-      2,
+      currency,
     );
     value = formattedValue;
-    displayValue = `$${formattedValue}`;
+    displayValue = formattedValue;
   }
 
   const { amount, unit } = getParsedAmount({
@@ -275,9 +281,9 @@ const getDisplayValues = (
     status: transaction.status,
     statusText: lodash.capitalize(transaction.status),
     displayValueWithoutUnit: value,
-    displayValue: isDiscreetMode ? '$****' : displayValue,
-    displayValueUnit: 'USD',
-    displayFeeValue: isDiscreetMode ? '$****' : displayFeeValue,
+    displayValue: isDiscreetMode ? '****' : displayValue,
+    displayValueUnit: currency.toUpperCase(),
+    displayFeeValue: isDiscreetMode ? '****' : displayFeeValue,
     displayFee: `${isDiscreetMode ? '****' : fee} ${feeUnit.abbr}`,
     amount: parseFloat(amount),
     displayFeeWithoutUnit: fee,
@@ -320,6 +326,7 @@ export const mapTransactionForDisplay = (params: {
   accounts: IAccount[];
   lang: ILangState;
   isDiscreetMode: boolean;
+  currency: string;
 }): TransactionRowData => {
   const { transaction } = params;
 
@@ -331,7 +338,12 @@ export const mapTransactionForDisplay = (params: {
   );
 
   return {
-    ...getDisplayValues(transaction, params.priceInfos, params.isDiscreetMode),
+    ...getDisplayValues(
+      transaction,
+      params.priceInfos,
+      params.isDiscreetMode,
+      params.currency,
+    ),
     id: transaction.__id ?? '',
     xpubOrAddress: account?.xpubOrAddress ?? '',
     walletAndAccount: getWalletAndAccount(wallet, account),
@@ -361,6 +373,7 @@ export const useTransactions = ({
   parentAssetId,
   accountId,
 }: UseTransactionsProps = {}) => {
+  const { currentCurrency } = useCurrency();
   const {
     lang,
     wallets,
@@ -368,7 +381,8 @@ export const useTransactions = ({
     transactions: allTransactions,
     priceInfos,
     isDiscreetMode,
-  } = useAppSelector(selector);
+  } = useAppSelector(state => selector(state, currentCurrency));
+
   const refData = useStateToRef({
     lang,
     wallets,
@@ -380,6 +394,7 @@ export const useTransactions = ({
     assetId,
     parentAssetId,
     accountId,
+    currentCurrency,
   });
 
   const dispatch = useAppDispatch();
@@ -470,6 +485,7 @@ export const useTransactions = ({
             wallets: refData.current.wallets,
             accounts: refData.current.accounts,
             lang: refData.current.lang,
+            currency: refData.current.currentCurrency,
           }),
         );
 
@@ -492,7 +508,15 @@ export const useTransactions = ({
 
   useEffect(() => {
     debounceParseTransactionListOnUserAction();
-  }, [isDiscreetMode, lang, walletId, assetId, parentAssetId, accountId]);
+  }, [
+    isDiscreetMode,
+    lang,
+    walletId,
+    assetId,
+    parentAssetId,
+    accountId,
+    currentCurrency,
+  ]);
 
   useEffect(() => {
     setDisplayedData(getDisplayDataList(transactionList));