+++ packages/cysync-core/src/pages/MainApp/hooks/useWalletPage.tsx
@@ -29,14 +29,15 @@ import {
 } from '~/actions';
 import { CoinIcon } from '~/components';
 import { routes } from '~/constants';
+import { useCurrency } from '~/context';
 import { useNavigateTo, useQuery } from '~/hooks';
 import {
   AccountSyncState,
   AccountSyncStateMap,
   selectAccountSync,
+  selectCurrentCurrencyPriceInfos,
   selectDiscreetMode,
   selectLanguage,
-  selectPriceInfos,
   selectUnHiddenAccounts,
   selectWallets,
   useAppDispatch,
@@ -127,6 +128,7 @@ const mapTokenAccounts = (
   a: IAccount,
   priceInfos: IPriceInfo[],
   isDiscreetMode: boolean,
+  currency: string,
 ): AccountTokenType => {
   const { amount, unit } = getParsedAmount({
     coinId: a.parentAssetId,
@@ -137,9 +139,7 @@ const mapTokenAccounts = (
 
   let displayValue = '$0.00';
   let value = '0.00';
-  const coinPrice = priceInfos.find(
-    p => p.assetId === a.assetId && p.currency.toLowerCase() === 'usd',
-  );
+  const coinPrice = priceInfos.find(p => p.assetId === a.assetId);
 
   if (coinPrice) {
     const balanceInDefaultUnit = convertToUnit({
@@ -153,9 +153,10 @@ const mapTokenAccounts = (
       new BigNumber(balanceInDefaultUnit.amount).multipliedBy(
         coinPrice.latestPrice,
       ),
+      currency,
     );
     value = formattedValue;
-    displayValue = `$${formattedValue}`;
+    displayValue = `${formattedValue}`;
   }
   const formattedAmount = formatDisplayAmount(amount, 24);
   const displayAmount = `${isDiscreetMode ? '****' : formattedAmount.fixed} ${
@@ -186,7 +187,7 @@ const mapTokenAccounts = (
     text: name,
     displayAmount,
     amountTooltip,
-    displayValue: isDiscreetMode ? '$****' : displayValue,
+    displayValue: isDiscreetMode ? '****' : displayValue,
     amount: parseFloat(amount),
     value: parseFloat(value),
   };
@@ -197,17 +198,19 @@ const selector = createSelector(
     selectLanguage,
     selectWallets,
     selectUnHiddenAccounts,
-    selectPriceInfos,
+    selectCurrentCurrencyPriceInfos,
     selectAccountSync,
     selectDiscreetMode,
+    (state, currency: string) => currency,
   ],
   (
     lang,
     { wallets },
     { accounts },
-    { priceInfos },
+    priceInfos,
     { accountSyncMap },
     { active: isDiscreetMode },
+    currency,
   ) => ({
     lang,
     wallets,
@@ -215,10 +218,12 @@ const selector = createSelector(
     priceInfos,
     accountSyncMap,
     isDiscreetMode,
+    currency,
   }),
 );
 
 export const useWalletPage = () => {
+  const { currentCurrency } = useCurrency();
   const {
     wallets,
     lang,
@@ -226,7 +231,7 @@ export const useWalletPage = () => {
     priceInfos,
     accountSyncMap,
     isDiscreetMode,
-  } = useAppSelector(selector);
+  } = useAppSelector(state => selector(state, currentCurrency));
   const dispatch = useAppDispatch();
   const navigateTo = useNavigateTo();
   const query = useQuery();
@@ -290,9 +295,7 @@ export const useWalletPage = () => {
 
       let displayValue = '$0.00';
       let value = '0.00';
-      const coinPrice = priceInfos.find(
-        p => p.assetId === a.assetId && p.currency.toLowerCase() === 'usd',
-      );
+      const coinPrice = priceInfos.find(p => p.assetId === a.assetId);
 
       if (coinPrice) {
         const balanceInDefaultUnit = convertToUnit({
@@ -306,15 +309,21 @@ export const useWalletPage = () => {
           new BigNumber(balanceInDefaultUnit.amount).multipliedBy(
             coinPrice.latestPrice,
           ),
+          currentCurrency,
         );
         value = formattedValue;
-        displayValue = `$${formattedValue}`;
+        displayValue = `${formattedValue}`;
       }
 
       const tokenAccounts = allAccounts
         .filter(ta => ta.parentAccountId === a.__id)
         .map(tokenAccount =>
-          mapTokenAccounts(tokenAccount, priceInfos, isDiscreetMode),
+          mapTokenAccounts(
+            tokenAccount,
+            priceInfos,
+            isDiscreetMode,
+            currentCurrency,
+          ),
         );
 
       if (tokenAccounts.length > 0) {
@@ -340,7 +349,7 @@ export const useWalletPage = () => {
           ] ?? checkComponent,
         displayAmount,
         amountTooltip,
-        displayValue: isDiscreetMode ? '$****' : displayValue,
+        displayValue: isDiscreetMode ? '****' : displayValue,
         amount: parseFloat(amount),
         value: parseFloat(value),
         account: a,
@@ -384,7 +393,9 @@ export const useWalletPage = () => {
     const account = accounts.find(a => a.__id === row.id);
 
     if (account) {
-      dispatch(syncAccounts({ accounts: [account] }));
+      dispatch(
+        syncAccounts({ accounts: [account], currency: currentCurrency }),
+      );
     }
   };